refactor!: new database schema and data models (#80)
* refactor: new database schema and data models

* feat: new BLOB column for image entries + `expires_at` column.
* perf: `CbEventContent` modified to store bytes instead of b64 encoded
string. encoding is only performed before sending data to frontend.
* perf: use microsecond UNIX timestamp for entry ids instead of relying
on sqlite autoincrement. allows to immediately send the event to
frontend instead of waiting on a db write.
* cleanup: remove `timestamp` column in db since no longer necessary.
* cleanup: get rid of `ClipboardEventNotice` since it is no longer
necessary. we're able to send the event immediately.
* style: comments and stuff.

* refactor: make `expires_at` nullable

it's absence means the entry never expires, such as when pinned.

* fix: DROP `_sqlx_migrations` instead of DELETE FROM

* fix: correct pinning

+++ src-tauri/src/config.rs
@@ -10,7 +10,7 @@ pub struct Config {
     pub theme: String,
 
     /// Number of days after which a clipboard history item is automatically deleted (unless pinned).
-    pub history_ttl: u32,
+    pub history_ttl: i64,
 
     /// Whether the clipboard history persists between app launches.
     pub persistence: bool,
